% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_king_tables.R
\name{check_king_tables}
\alias{check_king_tables}
\title{King Tables exists}
\usage{
check_king_tables(cfg)
}
\arguments{
\item{cfg}{king kongfig containing all tables to check}
}
\description{
King Tables exists
}
