% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_activity_data.R
\name{collect_act_df}
\alias{collect_act_df}
\title{collect activity data from BIG google}
\usage{
collect_act_df(cfg, df)
}
\arguments{
\item{cfg}{a King Kongfig}

\item{df}{a datafram containing the user ID needed for extraction of activity data}
}
\description{
collect activity data from BIG google
}
