% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_assignment_data.R
\name{collect_ass_df}
\alias{collect_ass_df}
\title{Collect assignment data from BigGoogle}
\usage{
collect_ass_df(cfg, rows)
}
\arguments{
\item{cfg}{a King Kongfig object}

\item{rows}{integer of wanted number of rows in query}
}
\description{
Collect assignment data from BigGoogle
}
