% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_collect.R
\name{data_tables}
\alias{data_tables}
\title{Collect data from Google BigQuery}
\usage{
data_tables(project, dataset, table, rows = NULL)
}
\arguments{
\item{project}{The google project which data will be collected from}

\item{dataset}{The data set from which the tables will be choosen}

\item{table}{The tables of choice from which one can colect data}

\item{rows}{Simple function for number of rows to extract, if NULL al rows will be collected}
}
\description{
Collect data from Google BigQuery
}
