% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/king_sql_query.R
\name{king_query}
\alias{king_query}
\title{King query}
\usage{
king_query(project, query)
}
\arguments{
\item{project}{the project from which data should be collected}

\item{query}{SQL query for calling BIGdata}
}
\description{
King query
}
