% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation.R
\name{model_eval}
\alias{model_eval}
\title{Model evaluation king}
\usage{
model_eval(king_df_list, king_model, predictions)
}
\arguments{
\item{king_df_list}{data frames started with}

\item{king_model}{model created}

\item{predictions}{predictions from test data set}
}
\description{
Model evaluation king
}
