% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_modell_df.R
\name{modelling_df}
\alias{modelling_df}
\title{Modelling data set}
\usage{
modelling_df(act_df, ass_df, train_prop)
}
\arguments{
\item{act_df}{activity data frame}

\item{ass_df}{assignment data frame}

\item{train_prop}{proportion of data to be used in the modelling}
}
\description{
Modelling data set
}
