% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/king_predict.R
\name{predict}
\alias{predict}
\title{Predictions for king}
\usage{
predict(king_model, prediction_data)
}
\arguments{
\item{king_model}{the models produced for the data}

\item{prediction_data}{the dataset which is going to be used in prediction}
}
\description{
Predictions for king
}
