% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_king_kongfig.R
\name{set_king_config}
\alias{set_king_config}
\title{King config assertion}
\usage{
set_king_config(args)
}
\arguments{
\item{args}{character vector to assert.}
}
\description{
King config assertion
}
