% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model.R
\name{train_model}
\alias{train_model}
\title{Train an arbitrary model}
\usage{
train_model(train_set, cfg)
}
\arguments{
\item{train_set}{a training data set}

\item{cfg}{a \code{king_kongfig} object.}
}
\description{
Train an arbitrary model
}
